Move validateAccountPageRoute into its own file
